PiPerformanceTestSuite.scala created

+++ src/main/scala/org/montecarlo/HasMultiplicity.scala
@@ -4,5 +4,5 @@ package org.montecarlo
  * Denotes a class that consists of a set of parameters/trials that can have multiple execution
  */
 trait HasMultiplicity {
-  def multiplicity(): Int
+  def multiplicity(): Long
 }
